pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "order-flow-trading"
version = "1.0.0"
description = "Real-time order flow analysis with regime-adaptive trading"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
databento = ["databento>=0.30.0"]
rithmic = []  # Add when Rithmic adapter is implemented
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "mypy>=1.4.0",
]
all = [
    "databento>=0.30.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "pandas-ta>=0.3.14b",
]

[project.scripts]
tradebot = "main:main"
tradebot-demo = "scripts.run_demo:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
